- migrate some 3d rendering away

+++ modapi/src/main/java/kr/syeyoung/modapi/rendering/URenderContext.java
@@ -6,5 +6,5 @@ public interface URenderContext {
     void popMatrix();
     void translate(double x, double y, double z);
     void scale(double x, double y, double z);
-    void rotate(float angle, int x, int y, int z);
+    void rotate(float angle, float x, float y, float z);
 }
